+++ massiv/src/Data/Massiv/Core/Common.hs
@@ -1,8 +1,11 @@
 {-# LANGUAGE BangPatterns          #-}
 {-# LANGUAGE DefaultSignatures     #-}
 {-# LANGUAGE FlexibleContexts      #-}
+{-# LANGUAGE MagicHash             #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE ScopedTypeVariables   #-}
 {-# LANGUAGE TypeFamilies          #-}
+{-# LANGUAGE UnboxedTuples         #-}
 {-# LANGUAGE UndecidableInstances  #-}
 -- |
 -- Module      : Data.Massiv.Core.Common
@@ -24,6 +27,8 @@ module Data.Massiv.Core.Common
   , InnerSlice(..)
   , Manifest(..)
   , Mutable(..)
+  , State(..)
+  , WorldState
   , Ragged(..)
   , Nested(..)
   , NestedStruct
@@ -44,10 +49,11 @@ module Data.Massiv.Core.Common
   , module Data.Massiv.Core.Computation
   ) where
 
-import           Control.Monad.Primitive      (PrimMonad (..))
+import           Control.Monad.Primitive
 import           Data.Massiv.Core.Computation
 import           Data.Massiv.Core.Index
 import           Data.Typeable
+import           GHC.Prim
 
 -- | The array family. Representations @r@ describes how data is arranged or computed. All arrays
 -- have a common property that each index @ix@ always maps to the same unique element, even if that
@@ -149,6 +155,11 @@ class Source r ix e => Manifest r ix e where
   unsafeLinearIndexM :: Array r ix e -> Int -> e
 
 
+data State s = State (State# s)
+
+type WorldState = State RealWorld
+
+
 class Manifest r ix e => Mutable r ix e where
   data MArray s r ix e :: *
 
@@ -176,6 +187,36 @@ class Manifest r ix e => Mutable r ix e where
 
   unsafeLinearWrite :: PrimMonad m =>
                        MArray (PrimState m) r ix e -> Int -> e -> m ()
+
+  -- | Create new mutable array, leaving it's elements uninitialized. Size isn't validated
+  -- either.
+  unsafeNewA :: Applicative f => ix -> WorldState -> f (WorldState, MArray RealWorld r ix e)
+  unsafeNewA sz (State s#) =
+    case internal (unsafeNew sz :: IO (MArray RealWorld r ix e)) s# of
+      (# s'#, ma #) -> pure (State s'#, ma)
+  {-# INLINE unsafeNewA #-}
+
+  unsafeThawA :: Applicative m =>
+                 Array r ix e -> WorldState -> m (WorldState, MArray RealWorld r ix e)
+  unsafeThawA arr (State s#) =
+    case internal (unsafeThaw arr :: IO (MArray RealWorld r ix e)) s# of
+      (# s'#, ma #) -> pure (State s'#, ma)
+  {-# INLINE unsafeThawA #-}
+
+  unsafeFreezeA :: Applicative m =>
+                   Comp -> MArray RealWorld r ix e -> WorldState -> m (WorldState, Array r ix e)
+  unsafeFreezeA comp marr (State s#) =
+    case internal (unsafeFreeze comp marr :: IO (Array r ix e)) s# of
+      (# s'#, a #) -> pure (State s'#, a)
+  {-# INLINE unsafeFreezeA #-}
+
+  unsafeLinearWriteA :: Applicative m =>
+                        MArray RealWorld r ix e -> Int -> e -> WorldState -> m WorldState
+  unsafeLinearWriteA marr i val (State s#) =
+    case internal (unsafeLinearWrite marr i val :: IO ()) s# of
+      (# s'#, _ #) -> pure (State s'#)
+  {-# INLINE unsafeLinearWriteA #-}
+
 
 
 class Nested r ix e where